reorganize typography code into separate files
