+++ packages/typography/src/Disclaimer.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { css, cx } from '@leafygreen-ui/emotion';
+import { HTMLElementProps } from '@leafygreen-ui/lib';
+import { sharedStyles } from './styles';
+
+/**
+ * Disclaimer
+ */
+const disclaimer = css`
+  display: block;
+  font-size: 12px;
+  line-height: 20px;
+  letter-spacing: 0px;
+`;
+
+type DisclaimerProps = HTMLElementProps<'small'>;
+
+function Disclaimer({ children, className, ...rest }: DisclaimerProps) {
+  return (
+    <small {...rest} className={cx(sharedStyles, disclaimer, className)}>
+      {children}
+    </small>
+  );
+}
+
+Disclaimer.displayName = 'Disclaimer';
+
+export default Disclaimer;
