--- packages/typography/src/FormElements/index.ts
@@ -1,2 +1,0 @@
-export { Label } from './Label';
-export { Description } from './Description';
