+++ packages/typography/src/Link.tsx
@@ -152,4 +152,4 @@ const Link: ExtendableBox<LinkProps, 'a'> = ({
   );
 };
 
-export { Link };
+export default Link;
