+++ packages/typography/src/index.ts
@@ -1,14 +1,13 @@
-export {
-  H1,
-  H2,
-  H3,
-  Subtitle,
-  Body,
-  InlineCode,
-  InlineKeyCode,
-  Disclaimer,
-  Overline,
-} from './Typography';
-export { Link } from './Link';
-export { Label, Description } from './FormElements';
 export { useUpdatedBaseFontSize } from './useUpdatedBaseFontSize';
+export { default as H1 } from './H1';
+export { default as H2 } from './H2';
+export { default as H3 } from './H3';
+export { default as Subtitle } from './Subtitle';
+export { default as Body } from './Body';
+export { default as Overline } from './Overline';
+export { default as InlineCode } from './InlineCode';
+export { default as InlineKeyCode } from './InlineKeyCode';
+export { default as Description } from './Description';
+export { default as Disclaimer } from './Disclaimer';
+export { default as Label } from './Label';
+export { default as Link } from './Link';
