+++ packages/typography/src/styles.tsx
@@ -1,4 +1,12 @@
 import { css } from '@leafygreen-ui/emotion';
+import { palette } from '@leafygreen-ui/palette';
+import { fontFamilies } from '@leafygreen-ui/tokens';
+
+export const sharedStyles = css`
+  margin: unset;
+  font-family: ${fontFamilies.default};
+  color: ${palette.black};
+`;
 
 // TODO: Refresh - Update this when DarkMode is designed
 export const typeScale1 = css`
